master-data-flask-app/requirements.txt
# Install with: pip install -r requirements.txt

flask
spacy
google-generativeai
# python -m spacy download en_core_web_sm